requirements.txt
Cython==0.29.14
numpy==1.16.5
absl-py==0.8.1
astor==0.8.0
backports.weakref==1.0.post1
cachetools==3.1.1
certifi==2019.9.11
chardet==3.0.4
Click==7.0
enum34==1.1.6
Flask==1.1.1
Flask-Cors==3.0.8
funcsigs==1.0.2
functools32==3.2.3.post2
futures==3.3.0
gast==0.2.2
google-auth==1.6.3
google-auth-oauthlib==0.4.1
google-pasta==0.1.7
grpcio==1.24.3
gunicorn==19.9.0
h5py==2.10.0
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.3
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
Markdown==3.1.1
MarkupSafe==1.1.1
mock==3.0.5
oauthlib==3.1.0
opencv-python==4.1.1.26
opt-einsum==2.3.2
Pillow==6.2.1
protobuf==3.10.0
pyasn1==0.4.7
pyasn1-modules==0.2.7
requests==2.22.0
requests-oauthlib==1.2.0
rsa==4.0
six==1.12.0
tensorboard==1.15.0
tensorflow==1.14.0
tensorflow-estimator==1.15.1
tensorflow-hub==0.6.0
termcolor==1.1.0
urllib3==1.25.6
Werkzeug==0.16.0
wget==3.2
wrapt==1.11.2
